src/requirements.txt
transformers
pillow
numpy
tqdm
timm
scikit-learn
matplotlib